pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "volc-speech-python-sdk"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "websockets>=14.0",
]
